Extract the sidebar header
The sidebar bundled its app chrome — window controls, project picker,
activity spinner, fetch and settings — with page switching and list
hosting. The chrome now has its own component.

- SidebarHeader is purely presentational; ActivitySpinner and
  FetchFromRemotesButton move with it, as do their styles
- The fetch/update wiring stays with the Sidebar, which feeds it to the
  header, the upstream list, and the hotkeys — the shared consumers are
  why it could not move into the header

No behavior change.
