+++ apps/lite/ui/src/routes/project/$id/workspace/Sidebar.tsx
@@ -7,98 +7,35 @@ import {
 	workspaceFetchStatusQueryOptions,
 } from "#ui/api/queries.ts";
 import { stackBottomRelativeTo } from "#ui/api/stack.ts";
-import { getButtonClassName } from "#ui/components/Button.tsx";
-import { classes } from "#ui/components/classes.ts";
 import { errorMessageForToast } from "#ui/errors.ts";
 import { Icon } from "#ui/components/Icon.tsx";
 import { TooltipPopup } from "#ui/components/Tooltip.tsx";
-import { globalHotkeys, workspaceHotkeys } from "#ui/hotkeys.ts";
+import { workspaceHotkeys } from "#ui/hotkeys.ts";
 import type { Address } from "#ui/addresses.ts";
 import { projectSlice } from "#ui/projects/state.ts";
 import { interfaceSlice } from "#ui/interface/state.ts";
 import { useAppDispatch, useAppSelector } from "#ui/store.ts";
-import { formatRelativeTime } from "#ui/time.ts";
 import type { AddressSpace } from "#ui/workspace/address-space.ts";
 import { Button, Toast, Toggle, ToggleGroup, Tooltip } from "@base-ui/react";
 import type { BottomUpdate, ProjectForFrontend } from "@gitbutler/but-sdk";
-import { LiteTestId } from "@gitbutler/ui/utils/testIds";
-import { useIsFetching, useIsMutating, useQuery } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import { useHotkeys } from "@tanstack/react-hotkeys";
-import { Match } from "effect";
-import { type FC, useRef, useState } from "react";
+import { type FC, useRef } from "react";
 import { ToggleGroupStyles, ToggleStyles } from "#ui/components/ToggleGroup.tsx";
 import { WorkspaceLists } from "#ui/routes/project/$id/workspace/WorkspaceLists/WorkspaceLists.tsx";
 import { BranchesList } from "#ui/routes/project/$id/workspace/BranchesList.tsx";
 import type { BranchesListData } from "#ui/routes/project/$id/workspace/useBranchesList.ts";
-import { FolderIcon } from "#ui/components/FolderIcon.tsx";
 import { UpstreamList } from "#ui/routes/project/$id/workspace/UpstreamList.tsx";
 import type { UpstreamListData } from "#ui/routes/project/$id/workspace/useUpstreamList.ts";
 import { assert } from "#ui/assert.ts";
 import { Badge } from "#ui/components/Badge.tsx";
 import type { PageId } from "#ui/projects/project.ts";
 import styles from "./Sidebar.module.css";
-import { TopLeftControls } from "#ui/routes/project/$id/workspace/TopLeftControls.tsx";
+import { SidebarHeader } from "#ui/routes/project/$id/workspace/SidebarHeader.tsx";
 import { useNewBranch } from "#ui/routes/project/$id/workspace/useNewBranch.ts";
 import { showNativeMenuFromTrigger } from "#ui/native-menu.ts";
 import { RowToolbar } from "#ui/routes/project/$id/workspace/Row.tsx";
 import { getRowButtonClassName } from "#ui/routes/project/$id/workspace/Row-utils.ts";
-
-const ActivitySpinner: FC<{
-	/** Suppressed while the fetch button shows its own spinner, to avoid two spinners at once. */
-	suppressed: boolean;
-}> = (p) => {
-	const fetchingCount = useIsFetching();
-	const mutatingCount = useIsMutating();
-
-	const isFetching = fetchingCount > 0;
-	const isMutating = mutatingCount > 0;
-
-	const status = Match.value({ isFetching, isMutating }).pipe(
-		Match.when({ isFetching: true, isMutating: true }, () => "Syncing"),
-		Match.when({ isFetching: true }, () => "Loading"),
-		Match.when({ isMutating: true }, () => "Saving"),
-		Match.orElse(() => null),
-	);
-
-	return !p.suppressed && status !== null && <Icon name="spinner" aria-label={status} />;
-};
-
-const FetchFromRemotesButton: FC<{
-	canFetch: boolean;
-	isPending: boolean;
-	lastSuccessfulMs?: number | null;
-	onFetch: () => void;
-}> = (p) => {
-	const [tooltipNow, setTooltipNow] = useState(() => Date.now());
-
-	return (
-		<Tooltip.Root
-			onOpenChange={(open) => {
-				if (open) setTooltipNow(Date.now());
-			}}
-		>
-			<Tooltip.Trigger
-				aria-label={workspaceHotkeys.fetchFromRemotes.meta.name}
-				className={getButtonClassName({ iconOnly: true, variant: "ghost" })}
-				onClick={p.onFetch}
-				// We pass `disabled` here because we want to disable the button, not
-				// the tooltip.
-				render={<Button focusableWhenDisabled disabled={!p.canFetch} />}
-			>
-				<Icon name={p.isPending ? "spinner" : "refresh"} />
-			</Tooltip.Trigger>
-			<Tooltip.Portal>
-				<Tooltip.Positioner sideOffset={4}>
-					<Tooltip.Popup render={<TooltipPopup kbd={workspaceHotkeys.fetchFromRemotes.hotkey} />}>
-						{workspaceHotkeys.fetchFromRemotes.meta.name}
-						{p.lastSuccessfulMs != null &&
-							` (${formatRelativeTime(p.lastSuccessfulMs, tooltipNow)})`}
-					</Tooltip.Popup>
-				</Tooltip.Positioner>
-			</Tooltip.Portal>
-		</Tooltip.Root>
-	);
-};
 
 /** The tabs in the order they are shown, for cycling with `[` and `]`. */
 const pageOrder: Array<PageId> = ["workspace", "upstream", "branches"];
@@ -199,10 +136,6 @@ export const Sidebar: FC<{
 
 	const canCreateBranch = newBranch.enabled;
 
-	const canApplyBranch = noOperationPending;
-
-	const canOpenSettings = noOperationPending;
-
 	const ref = useRef<HTMLDivElement>(null);
 
 	useHotkeys([
@@ -212,7 +145,7 @@ export const Sidebar: FC<{
 			options: {
 				conflictBehavior: "allow",
 				meta: workspaceHotkeys.applyBranch.meta,
-				enabled: canApplyBranch,
+				enabled: noOperationPending,
 			},
 		},
 		{
@@ -280,65 +213,16 @@ export const Sidebar: FC<{
 	return (
 		<div className={styles.container} ref={ref}>
 			<div className={styles.top}>
-				<header className={styles.workspaceControls}>
-					<TopLeftControls />
-
-					<div className={styles.workspaceControlsLeft}>
-						<Tooltip.Root>
-							<Tooltip.Trigger
-								aria-label={`${globalHotkeys.selectProject.meta.name} (current: ${project.title})`}
-								data-testid={LiteTestId.ProjectPickerButton}
-								className={classes(
-									getButtonClassName({ variant: "ghost" }),
-									"text-15",
-									"text-bold",
-									styles.workspaceName,
-								)}
-								onClick={openProjectPicker}
-							>
-								<FolderIcon className={styles.workspaceNameFolder} />
-								<span className={styles.workspaceNameLabel}>{project.title}</span>
-							</Tooltip.Trigger>
-							<Tooltip.Portal>
-								<Tooltip.Positioner sideOffset={4}>
-									<Tooltip.Popup render={<TooltipPopup kbd={globalHotkeys.selectProject.hotkey} />}>
-										{globalHotkeys.selectProject.meta.name}
-									</Tooltip.Popup>
-								</Tooltip.Positioner>
-							</Tooltip.Portal>
-						</Tooltip.Root>
-						<ActivitySpinner suppressed={isWorkspaceFetchFromRemotesPending} />
-					</div>
-
-					<div className={styles.workspaceControlsActions}>
-						<FetchFromRemotesButton
-							canFetch={canFetchFromRemotes}
-							isPending={isWorkspaceFetchFromRemotesPending}
-							lastSuccessfulMs={workspaceFetchStatus?.lastSuccessfulMs}
-							onFetch={fetchFromRemotes}
-						/>
-
-						<Tooltip.Root>
-							<Tooltip.Trigger
-								aria-label={workspaceHotkeys.settings.meta.name}
-								className={getButtonClassName({ iconOnly: true, variant: "ghost" })}
-								onClick={openSettings}
-								// We pass `disabled` here because we want to disable the button, not
-								// the tooltip. Other props should be passed above.
-								render={<Button focusableWhenDisabled disabled={!canOpenSettings} />}
-							>
-								<Icon name="settings" />
-							</Tooltip.Trigger>
-							<Tooltip.Portal>
-								<Tooltip.Positioner sideOffset={4}>
-									<Tooltip.Popup render={<TooltipPopup kbd={workspaceHotkeys.settings.hotkey} />}>
-										{workspaceHotkeys.settings.meta.name}
-									</Tooltip.Popup>
-								</Tooltip.Positioner>
-							</Tooltip.Portal>
-						</Tooltip.Root>
-					</div>
-				</header>
+				<SidebarHeader
+					project={project}
+					canFetch={canFetchFromRemotes}
+					isFetchPending={isWorkspaceFetchFromRemotesPending}
+					lastSuccessfulFetchMs={workspaceFetchStatus?.lastSuccessfulMs}
+					onFetch={fetchFromRemotes}
+					canOpenSettings={noOperationPending}
+					onOpenSettings={openSettings}
+					onOpenProjectPicker={openProjectPicker}
+				/>
 
 				<ToggleGroup
 					render={<ToggleGroupStyles />}
